requirements-test.txt
coveralls
django-extensions>=2.0.7
openwisp-utils[qa]>=0.4
